// packages/web/src/generated/rust-types/SshConnectionEntry.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type SshConnectionEntry = { 
/**
 * Protocol (tcp, udp)
 */
protocol: string, 
/**
 * Local address
 */
localAddress: string, 
/**
 * Peer/remote address
 */
peerAddress: string, 
/**
 * Connection state (ESTAB, TIME-WAIT, etc.)
 */
state: string, 
/**
 * Process name
 */
process: string, };