// packages/web/src/generated/rust-types/SshListeningSocketEntry.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type SshListeningSocketEntry = { 
/**
 * Protocol (tcp, udp)
 */
protocol: string, 
/**
 * Local address (e.g. 0.0.0.0:80, :::443)
 */
localAddress: string, 
/**
 * Peer address
 */
peerAddress: string, 
/**
 * Socket state (LISTEN, ESTAB, etc.)
 */
state: string, 
/**
 * Process info (e.g. "nginx" or "pid/name")
 */
process: string, };