// packages/web/src/generated/rust-types/SshNetworkInterfaceEntry.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type SshNetworkInterfaceEntry = { 
/**
 * Interface name (e.g. eth0, lo)
 */
name: string, 
/**
 * IP addresses with prefix (e.g. 192.168.1.2/24)
 */
ipAddresses: Array<string>, 
/**
 * MAC address (e.g. 00:11:22:33:44:55)
 */
macAddress: string, 
/**
 * Whether the interface is up
 */
isUp: boolean, 
/**
 * MTU value
 */
mtu: number | null, 
/**
 * Total received bytes
 */
rxBytes: number, 
/**
 * Total transmitted bytes
 */
txBytes: number, };