// packages/web/src/generated/rust-types/SshRouteEntry.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type SshRouteEntry = { 
/**
 * Destination network (e.g. "default", "192.168.1.0/24")
 */
destination: string, 
/**
 * Gateway address (e.g. "172.17.0.1")
 */
gateway: string, 
/**
 * Network interface (e.g. "eth0")
 */
iface: string, 
/**
 * Protocol (e.g. "kernel", "dhcp", "static")
 */
protocol: string, 
/**
 * Scope (e.g. "link", "global")
 */
scope: string, 
/**
 * Metric value
 */
metric: string, };